new scatterplots
I went ahead and added in the scatterplots she was interested in seeing. After generating the visualizations, it was clear that no significant or noticeable relationships were coming from the scatterplots. I'm going to continue experimenting with the column information and work from there in hopes of creating new meaningful visualizations.